requirements.txt
django
djangorestframework
django-crispy-forms
numpy 
pandas 
sklearn 
pickle5
whitenoise
isort
joblib